renamed boos to boss-2
